replaced @autowired with @RequiredArgsConstructor, changed RSSItem ID from pubDate to generated ID, set permalink unique, simplified UpdateService, moved DB config to env variables

+++ src/main/java/cz/RSS/archive/springbootcrudjpaPostgre/repository/ItemRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @RepositoryRestResource
-public interface ItemRepository extends JpaRepository<RSSItem, Integer> {
+public interface ItemRepository extends JpaRepository<RSSItem, Long> {
     Optional<RSSItem> findFirstByStreamIdOrderByPubDateDesc(int streamId);
     List<RSSItem> findByStreamIdInOrderByPubDateDesc(List<Integer> streamId);
